pyproject.toml
[tool.poetry]
name = "mutli-agent-concierge"
version = "0.1.0"
description = ""
authors = ["Laurie Voss <github@seldo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index-core = "^0.10.52.post2"
llama-index-llms-openai = "^0.1.25"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
llama-index-llms-anthropic = "^0.1.17"
llama-index-agent-openai = "^0.2.9"
llama-index-utils-workflow = "^0.1.1"
pytrends = "^4.9.2"
yfinance = "^0.2.43"
nltk = "^3.9.1"
statistics = "^1.0.3.5"
tweepy = "^4.14.0"
textblob = "^0.18.0.post0"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
